requirements.txt
numpy==1.26.4
matplotlib==3.9.0
jupyter==1.0.0
openai==0.28.0
PyYAML==6.0.1